1. etl流程框架完成 2. mysqlutil 3. model和dao的封装

+++ src/main/scala/com/yejunyu/etl/model/DayVideoAccessStat.scala
@@ -4,4 +4,4 @@ package com.yejunyu.etl.model
   * @author yejunyu
   * @date 18-7-31. 
   */
-case class DayVideoAccessStat(day: String, cmsId: Long, times: Long)
+case class DayVideoAccessStat(day: String, cmsId: Int, times: Long)
